# Perf: faster final exponentiation (hard part) for BW6 curves