# Description

This PR otpimizes the final exponentiation (hard part) of BW6 curves.
Fixes #457.

Todo:
- [x] BW6-761
- [x] BW6-756
- [x] BW6-633

The PR also removes some unnecessary computations in the Miller loop (some cases do not occur given the loopCounters).
## Type of change

Please delete options that are not relevant.

- [x] New feature (non-breaking change which adds functionality)

# How has this been tested?

Existing pairing tests still apply to the new version.

# How has this been benchmarked?

- [x] Benchmark A, on Macbook Air M1, 16GB RAM
- BW6-761
```
benchmark              old ns/op     new ns/op     delta
BenchmarkPairing-8     2592580       2343537       -9.61%
```
- BW6-756
```
benchmark              old ns/op     new ns/op     delta
BenchmarkPairing-8     2566568       2353209       -8.31%
```
- BW6-633
```
benchmark              old ns/op     new ns/op     delta
BenchmarkPairing-8     1574459       1463099       -7.07%
```
- [x] Benchmark B, on AWS z1d.large (Intel(R) Xeon(R) Platinum 8151 CPU @ 3.40GHz)
- BW6-761
```
benchmark              old ns/op     new ns/op     delta
BenchmarkPairing-2     2250893       2099406       -6.73%
```
- BW6-756
```
benchmark              old ns/op     new ns/op     delta
BenchmarkPairing-2     2453394       2321157       -5.39%
```
- BW6-633
```
benchmark              old ns/op     new ns/op     delta
BenchmarkPairing-2     1579874       1478325       -6.43%
```

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] I did not modify files generated from templates
- [x] `golangci-lint` does not output errors locally
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream modules