@yelhousni 2023-10-16
@ivokub I just added BW6-756 and BW6-633 new final exponentiations. For BW6-756 it's the same algorithm as the BW6-761 you reviewed (case t % r % x₀ = 3 in BW6-BLS12 family and algorithm 4.4 from [[PhD:Housni22]](https://yelhousni.github.io/phd.pdf)). It just happens that `Expc1` is a `Conjugate` and `Expc2` is the identity map. For BW6-633, it is different (case t % r % x₀ = 0 in BW6-BLS24 family and algorithm 4.5).